Remove old provider
