+++ packages/discovery/src/discovery/provider/AllProviders.ts
@@ -4,12 +4,12 @@ import { DiscoveryChainConfig } from '../../config/types'
 import { EtherscanLikeClient } from '../../utils/EtherscanLikeClient'
 import { HttpClient } from '../../utils/HttpClient'
 import { BatchingAndCachingProvider } from './BatchingAndCachingProvider'
-import { getBlockNumberTwoProviders } from './DiscoveryProvider'
+import { DiscoveryCache } from './DiscoveryCache'
 import { HighLevelProvider } from './HighLevelProvider'
 import { IProvider, RawProviders } from './IProvider'
 import { LowLevelProvider } from './LowLevelProvider'
-import { DiscoveryCache } from './ProviderWithCache'
 import { ReorgAwareCache } from './ReorgAwareCache'
+import { getBlockNumberTwoProviders } from './getBlockNumberTwoProviders'
 import { MulticallClient } from './multicall/MulticallClient'
 
 export class AllProviders {
