+++ packages/discovery/src/discovery/provider/DiscoveryCache.ts
@@ -1,0 +1,9 @@
+export interface DiscoveryCache {
+  set(
+    key: string,
+    value: string,
+    chain: string,
+    blockNumber: number,
+  ): Promise<void>
+  get(key: string): Promise<string | undefined>
+}
