+++ packages/discovery/src/index.ts
@@ -30,12 +30,10 @@ export {
   discoveryDiffToMarkdown,
 } from './discovery/output/diffToMarkdown'
 export { toDiscoveryOutput } from './discovery/output/toDiscoveryOutput'
-export { DiscoveryProvider } from './discovery/provider/DiscoveryProvider'
 export { MulticallClient } from './discovery/provider/multicall/MulticallClient'
 export { getMulticall3Config } from './discovery/provider/multicall/MulticallConfig'
 export type { MulticallConfig } from './discovery/provider/multicall/types'
-export type { DiscoveryCache } from './discovery/provider/ProviderWithCache'
-export { ProviderWithCache } from './discovery/provider/ProviderWithCache'
+export type { DiscoveryCache } from './discovery/provider/DiscoveryCache'
 export { RateLimitedProvider } from './discovery/provider/RateLimitedProvider'
 export { ProxyDetector } from './discovery/proxies/ProxyDetector'
 export { deduplicateAbi } from './discovery/source/deduplicateAbi'
